refactor(server): split tool_risk into a package
The get_risk MCP tool had grown to ~1034 lines: single-target risk
scoring, post-assessment result enrichment, and PR-mode directive
assembly (breaking-change, cross-repo, conformance, governance) all in
one module alongside the tool entrypoint.

Split it into a tool_risk/ package:

- assessment.py   single-target risk scoring helpers
- enrichment.py   cross-repo / health / dep-summary result enrichment
- directives.py   PR-mode directive assembly
- get_risk.py     the @mcp.tool() orchestrator

Pure structural move: every symbol relocates verbatim, lazy imports stay
function-local, and absolute sibling imports are preserved. The
@mcp.tool() decorator on get_risk stays in get_risk.py and __init__.py
imports it, so import-time registration fires exactly once. __init__.py
re-exports get_risk plus the helpers routers/git.py and the tests import,
so the public path repowise.server.mcp_server.tool_risk is unchanged.
