+++ packages/server/src/repowise/server/mcp_server/tool_risk/__init__.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from .assessment import _check_test_gap, _classify_risk_type  # noqa: F401
+from .directives import (  # noqa: F401
+    _breaking_change_directive,
+    _conformance_directive,
+    _cross_repo_directive,
+    _trim_blast_lists,
+)
+from .get_risk import get_risk
+
+__all__ = ["get_risk"]
